wip(#1101): extract lower/closure_analysis
